new attempt at fixing bags
